feat: you-goal-p21-interrupt-active-dispatch-004 - Prove the Interrupt and Late-Result Race Behavior
