+++ pkg/api/servertests/server_factory_preview_test.go
@@ -5,12 +5,35 @@ import (
 	"encoding/json"
 	"net/http"
 	"net/http/httptest"
+	"os"
+	"path/filepath"
 	"strings"
 	"testing"
 
 	factoryapi "github.com/portpowered/infinite-you/pkg/api/generated"
 	workflowsource "github.com/portpowered/infinite-you/pkg/orchestrators/javascript/source"
 )
+
+const validWorkflowPreviewSource = `
+meta({ name: "review", version: 1 });
+phase("setup");
+log("starting");
+`
+
+func writeWorkflowPreviewFixture(t *testing.T, projectRoot, name, content string) {
+	t.Helper()
+	workflowDir := filepath.Join(projectRoot, workflowsource.ProjectClaudeWorkflowsDir)
+	if err := os.MkdirAll(workflowDir, 0o755); err != nil {
+		t.Fatalf("mkdir workflows: %v", err)
+	}
+	if err := os.WriteFile(filepath.Join(workflowDir, name), []byte(content), 0o600); err != nil {
+		t.Fatalf("write workflow: %v", err)
+	}
+}
+
+func stringPtr(value string) *string {
+	return &value
+}
 
 func TestPreviewFactory_ReturnsCanonicalFactoryPreviewContract(t *testing.T) {
 	projectRoot := t.TempDir()
