+++ pkg/factorysessionexecution/lifecycle.go
@@ -1,8 +1,12 @@
 package factorysessionexecution
 
 import (
+	"encoding/json"
+	"fmt"
 	"strings"
 	"time"
+
+	factoryapi "github.com/portpowered/infinite-you/pkg/api/generated"
 )
 
 // ProjectedLifecycleControlStatus returns the lifecycle-control status that live
@@ -204,12 +208,12 @@ const (
 type LifecycleControlKind string
 
 const (
-	LifecycleControlPause         LifecycleControlKind = "PAUSE"
-	LifecycleControlResume        LifecycleControlKind = "RESUME"
-	LifecycleControlCancel        LifecycleControlKind = "CANCEL"
-	LifecycleControlTerminate     LifecycleControlKind = "TERMINATE"
-	LifecycleControlApprove       LifecycleControlKind = "APPROVE"
-	LifecycleControlRetryDispatch    LifecycleControlKind = "RETRY_DISPATCH"
+	LifecycleControlPause             LifecycleControlKind = "PAUSE"
+	LifecycleControlResume            LifecycleControlKind = "RESUME"
+	LifecycleControlCancel            LifecycleControlKind = "CANCEL"
+	LifecycleControlTerminate         LifecycleControlKind = "TERMINATE"
+	LifecycleControlApprove           LifecycleControlKind = "APPROVE"
+	LifecycleControlRetryDispatch     LifecycleControlKind = "RETRY_DISPATCH"
 	LifecycleControlInterruptDispatch LifecycleControlKind = "INTERRUPT_DISPATCH"
 )
 
@@ -258,6 +262,449 @@ type PhaseSummary struct {
 	DispatchCount          int
 	CompletedDispatchCount int
 	FailedDispatchCount    int
+}
+
+const (
+	defaultDispatchInterruptionReason     = "Operator interrupted active dispatch"
+	dispatchInterruptionFailureReasonCode = "DISPATCH_INTERRUPTED"
+)
+
+type dispatchInterruptedEventPayload struct {
+	Reason         string `json:"reason"`
+	ObservedStatus string `json:"observedStatus"`
+	InterruptedAt  string `json:"interruptedAt"`
+	RetryPlanned   bool   `json:"retryPlanned"`
+}
+
+type interruptedDispatchPreservation struct {
+	summary              DispatchSummary
+	statusTransitions    []DispatchStatus
+	javaScriptProjection *DispatchJavaScriptProjection
+}
+
+// ObservedCancellationStatusForInterrupt returns the provider or process cancellation
+// status observed when one dispatch interruption is recorded.
+func ObservedCancellationStatusForInterrupt(priorStatus DispatchStatus) factoryapi.FactoryDispatchStatus {
+	switch priorStatus {
+	case DispatchStatusRunning:
+		return factoryapi.FactoryDispatchStatusRUNNING
+	default:
+		return factoryapi.FactoryDispatchStatusINTERRUPTED
+	}
+}
+
+func dispatchInterruptionReason(interrupt InterruptDispatchRequest) string {
+	if reason := strings.TrimSpace(interrupt.Reason); reason != "" {
+		return reason
+	}
+	return defaultDispatchInterruptionReason
+}
+
+func dispatchInterruptionFailureDetail(reason string) *DispatchFailureDetail {
+	return &DispatchFailureDetail{
+		Reason:  dispatchInterruptionFailureReasonCode,
+		Message: reason,
+	}
+}
+
+// MarkDispatchInterrupted updates one dispatch summary and transition history for interruption.
+func MarkDispatchInterrupted(
+	dispatches []DispatchSummary,
+	dispatchStatusTransitions map[string][]DispatchStatus,
+	dispatchID string,
+	interrupt InterruptDispatchRequest,
+) ([]DispatchSummary, map[string][]DispatchStatus) {
+	reason := dispatchInterruptionReason(interrupt)
+	failureDetail := dispatchInterruptionFailureDetail(reason)
+	for index, dispatch := range dispatches {
+		if dispatch.ID != dispatchID {
+			continue
+		}
+		dispatch.Status = DispatchStatusInterrupted
+		dispatch.FailureDetail = failureDetail
+		dispatches[index] = dispatch
+		if dispatchStatusTransitions != nil {
+			if transitions, ok := dispatchStatusTransitions[dispatchID]; ok {
+				dispatchStatusTransitions[dispatchID] = append(transitions, DispatchStatusInterrupted)
+			} else {
+				dispatchStatusTransitions[dispatchID] = []DispatchStatus{DispatchStatusInterrupted}
+			}
+		}
+		break
+	}
+	return dispatches, dispatchStatusTransitions
+}
+
+// AppendDispatchInterruptedEvent appends one canonical DISPATCH_INTERRUPTED event.
+func AppendDispatchInterruptedEvent(
+	events []json.RawMessage,
+	session SessionReadResult,
+	dispatch DispatchSummary,
+	interrupt InterruptDispatchRequest,
+	priorStatus DispatchStatus,
+	source string,
+) []json.RawMessage {
+	if strings.TrimSpace(session.SessionID) == "" || strings.TrimSpace(dispatch.ID) == "" {
+		return events
+	}
+	interruptedAt := time.Now().UTC()
+	reason := dispatchInterruptionReason(interrupt)
+	observedStatus := ObservedCancellationStatusForInterrupt(priorStatus)
+	sequence, sessionSequence := nextCanonicalEventSequence(events)
+	eventTime := canonicalSessionEventTime(session).Add(time.Duration(sessionSequence) * time.Second)
+	if interruptedAt.After(eventTime) {
+		eventTime = interruptedAt
+	}
+
+	sessionID := session.SessionID
+	orchestratorKind := strings.ToUpper(strings.TrimSpace(session.OrchestratorKind))
+	var orchestratorDialect *string
+	if dialect := strings.TrimSpace(session.Dialect); dialect != "" {
+		orchestratorDialect = &dialect
+	}
+	var phaseID *string
+	var phaseName *string
+	if phase := strings.TrimSpace(dispatch.Phase); phase != "" {
+		phaseID = &phase
+		phaseName = &phase
+	} else if phase := strings.TrimSpace(session.Phase); phase != "" {
+		phaseID = &phase
+		phaseName = &phase
+	}
+	dispatchID := dispatch.ID
+
+	payload, err := json.Marshal(dispatchInterruptedEventPayload{
+		Reason:         reason,
+		ObservedStatus: string(observedStatus),
+		InterruptedAt:  interruptedAt.Format(time.RFC3339),
+		RetryPlanned:   false,
+	})
+	if err != nil {
+		return events
+	}
+
+	context := canonicalFactoryEventContext{
+		Sequence:        sequence,
+		Tick:            sequence,
+		EventTime:       eventTime,
+		SessionID:       &sessionID,
+		SessionSequence: intPtr(sessionSequence),
+		Source:          &source,
+		DispatchID:      &dispatchID,
+	}
+	if orchestratorKind != "" {
+		context.OrchestratorKind = &orchestratorKind
+	}
+	if orchestratorDialect != nil {
+		context.OrchestratorDialect = orchestratorDialect
+	}
+	if phaseID != nil {
+		context.PhaseID = phaseID
+	}
+	if phaseName != nil {
+		context.PhaseName = phaseName
+	}
+
+	encoded, err := json.Marshal(canonicalFactoryEvent{
+		SchemaVersion: canonicalFactoryEventSchemaVersion,
+		ID:            fmt.Sprintf("dispatch-interrupted/%s/%d", dispatchID, sequence),
+		Type:          "DISPATCH_INTERRUPTED",
+		Context:       context,
+		Payload:       payload,
+	})
+	if err != nil {
+		return events
+	}
+	return append(events, encoded)
+}
+
+func nextCanonicalEventSequence(events []json.RawMessage) (sequence int, sessionSequence int) {
+	maxSequence := 0
+	maxSessionSequence := -1
+	for _, raw := range events {
+		parsed, err := parseCanonicalEvent(raw)
+		if err != nil {
+			continue
+		}
+		if parsed.Sequence > maxSequence {
+			maxSequence = parsed.Sequence
+		}
+		if parsed.SessionSequence != nil && *parsed.SessionSequence > maxSessionSequence {
+			maxSessionSequence = *parsed.SessionSequence
+		}
+	}
+	return maxSequence + 1, maxSessionSequence + 1
+}
+
+// ReplayDispatchProjection reconstructs durable dispatch summaries from canonical session events.
+func ReplayDispatchProjection(events []json.RawMessage) ([]DispatchSummary, error) {
+	dispatches := make(map[string]DispatchSummary)
+	for index, raw := range events {
+		if err := applyDispatchProjectionEvent(dispatches, raw); err != nil {
+			return nil, fmt.Errorf("apply event %d: %w", index, err)
+		}
+	}
+	if len(dispatches) == 0 {
+		return nil, nil
+	}
+	ordered := make([]DispatchSummary, 0, len(dispatches))
+	for _, dispatch := range dispatches {
+		ordered = append(ordered, dispatch)
+	}
+	return ordered, nil
+}
+
+func applyDispatchProjectionEvent(dispatches map[string]DispatchSummary, raw json.RawMessage) error {
+	var envelope canonicalFactoryEvent
+	if err := json.Unmarshal(raw, &envelope); err != nil {
+		return fmt.Errorf("unmarshal event envelope: %w", err)
+	}
+	switch strings.TrimSpace(envelope.Type) {
+	case "DISPATCH_INTERRUPTED":
+		return applyDispatchInterruptedProjection(dispatches, envelope)
+	default:
+		return nil
+	}
+}
+
+func applyDispatchInterruptedProjection(dispatches map[string]DispatchSummary, envelope canonicalFactoryEvent) error {
+	dispatchID := stringValuePtr(envelope.Context.DispatchID)
+	if dispatchID == "" {
+		return fmt.Errorf("DISPATCH_INTERRUPTED missing dispatchId")
+	}
+	var payload dispatchInterruptedEventPayload
+	if err := json.Unmarshal(envelope.Payload, &payload); err != nil {
+		return fmt.Errorf("unmarshal DISPATCH_INTERRUPTED payload: %w", err)
+	}
+	reason := strings.TrimSpace(payload.Reason)
+	if reason == "" {
+		reason = defaultDispatchInterruptionReason
+	}
+	dispatches[dispatchID] = DispatchSummary{
+		ID:     dispatchID,
+		Status: DispatchStatusInterrupted,
+		Phase:  stringValuePtr(envelope.Context.PhaseID),
+		FailureDetail: &DispatchFailureDetail{
+			Reason:  dispatchInterruptionFailureReasonCode,
+			Message: reason,
+		},
+	}
+	return nil
+}
+
+func snapshotInterruptedDispatches(state *runtimeSessionState) map[string]interruptedDispatchPreservation {
+	if state == nil || len(state.dispatches) == 0 {
+		return nil
+	}
+	preserved := make(map[string]interruptedDispatchPreservation)
+	for _, dispatch := range state.dispatches {
+		if dispatch.Status != DispatchStatusInterrupted {
+			continue
+		}
+		preservation := interruptedDispatchPreservation{
+			summary:           cloneDispatchSummary(dispatch),
+			statusTransitions: cloneDispatchStatusSlice(state.dispatchStatusTransitions[dispatch.ID]),
+		}
+		if js, ok := state.dispatchJavaScript[dispatch.ID]; ok {
+			projection := js
+			preservation.javaScriptProjection = &projection
+		}
+		preserved[dispatch.ID] = preservation
+	}
+	if len(preserved) == 0 {
+		return nil
+	}
+	return preserved
+}
+
+func restoreInterruptedDispatchResultSuppression(
+	state *runtimeSessionState,
+	preserved map[string]interruptedDispatchPreservation,
+) {
+	if state == nil || len(preserved) == 0 {
+		return
+	}
+	projectedByID := make(map[string]DispatchSummary, len(state.dispatches))
+	for _, dispatch := range state.dispatches {
+		projectedByID[dispatch.ID] = dispatch
+	}
+	for index, dispatch := range state.dispatches {
+		preservation, ok := preserved[dispatch.ID]
+		if !ok {
+			continue
+		}
+		state.dispatches[index] = enrichInterruptedDispatchDiagnostics(
+			preservation.summary,
+			projectedByID[dispatch.ID],
+		)
+	}
+	for dispatchID, preservation := range preserved {
+		if _, ok := projectedByID[dispatchID]; ok {
+			continue
+		}
+		state.dispatches = append(state.dispatches, cloneDispatchSummary(preservation.summary))
+	}
+	if state.dispatchStatusTransitions == nil {
+		state.dispatchStatusTransitions = make(map[string][]DispatchStatus, len(preserved))
+	}
+	for dispatchID, preservation := range preserved {
+		if len(preservation.statusTransitions) > 0 {
+			state.dispatchStatusTransitions[dispatchID] = cloneDispatchStatusSlice(preservation.statusTransitions)
+		}
+		if preservation.javaScriptProjection != nil {
+			if state.dispatchJavaScript == nil {
+				state.dispatchJavaScript = make(map[string]DispatchJavaScriptProjection)
+			}
+			state.dispatchJavaScript[dispatchID] = *preservation.javaScriptProjection
+		}
+	}
+	state.artifacts = filterArtifactsSuppressingInterruptedLateResults(state.artifacts, preserved)
+	recalculateSessionProgressFromDispatches(state)
+}
+
+func enrichInterruptedDispatchDiagnostics(preserved, projected DispatchSummary) DispatchSummary {
+	enriched := cloneDispatchSummary(preserved)
+	if enriched.Label == "" {
+		enriched.Label = projected.Label
+	}
+	if enriched.RunnerID == "" {
+		enriched.RunnerID = projected.RunnerID
+	}
+	if enriched.Model == "" {
+		enriched.Model = projected.Model
+	}
+	if enriched.Provider == "" {
+		enriched.Provider = projected.Provider
+	}
+	if len(enriched.ProviderSessionRefs) == 0 && len(projected.ProviderSessionRefs) > 0 {
+		enriched.ProviderSessionRefs = cloneProviderSessionRefs(projected.ProviderSessionRefs)
+	}
+	return enriched
+}
+
+func filterArtifactsSuppressingInterruptedLateResults(
+	artifacts []ArtifactSummary,
+	preserved map[string]interruptedDispatchPreservation,
+) []ArtifactSummary {
+	if len(artifacts) == 0 || len(preserved) == 0 {
+		return artifacts
+	}
+	filtered := make([]ArtifactSummary, 0, len(artifacts))
+	for _, artifact := range artifacts {
+		dispatchID := strings.TrimSpace(artifact.DispatchID)
+		if dispatchID == "" {
+			filtered = append(filtered, artifact)
+			continue
+		}
+		if _, interrupted := preserved[dispatchID]; interrupted && artifact.Kind == "CHILD_OUTPUT" {
+			continue
+		}
+		filtered = append(filtered, artifact)
+	}
+	return filtered
+}
+
+func recalculateSessionProgressFromDispatches(state *runtimeSessionState) {
+	if state == nil {
+		return
+	}
+	phaseCount := 0
+	if state.session.Progress != nil {
+		phaseCount = state.session.Progress.PhaseCount
+	}
+	progress := progressCountsFromDispatches(state.dispatches, phaseCount)
+	state.session.Progress = &progress
+	state.session.ArtifactRefs = artifactRefsFromSummaries(state.artifacts)
+	state.session.ArtifactCount = len(state.session.ArtifactRefs)
+}
+
+func extractDispatchInterruptedEvents(events []json.RawMessage) []json.RawMessage {
+	if len(events) == 0 {
+		return nil
+	}
+	interrupted := make([]json.RawMessage, 0)
+	for _, raw := range events {
+		var envelope factoryEventEnvelope
+		if err := json.Unmarshal(raw, &envelope); err != nil {
+			continue
+		}
+		if strings.TrimSpace(envelope.Type) != "DISPATCH_INTERRUPTED" {
+			continue
+		}
+		interrupted = append(interrupted, append(json.RawMessage(nil), raw...))
+	}
+	return interrupted
+}
+
+func mergePreservedDispatchInterruptedEvents(projected, preserved []json.RawMessage) []json.RawMessage {
+	if len(preserved) == 0 {
+		return projected
+	}
+	merged := make([]json.RawMessage, 0, len(projected)+len(preserved))
+	seen := make(map[string]struct{}, len(preserved))
+	for _, raw := range projected {
+		var envelope factoryEventEnvelope
+		if err := json.Unmarshal(raw, &envelope); err != nil {
+			merged = append(merged, raw)
+			continue
+		}
+		if strings.TrimSpace(envelope.Type) == "DISPATCH_INTERRUPTED" {
+			seen[eventIdentityKey(raw)] = struct{}{}
+		}
+		merged = append(merged, raw)
+	}
+	for _, raw := range preserved {
+		key := eventIdentityKey(raw)
+		if _, ok := seen[key]; ok {
+			continue
+		}
+		merged = append(merged, raw)
+		seen[key] = struct{}{}
+	}
+	return merged
+}
+
+func eventIdentityKey(raw json.RawMessage) string {
+	var envelope struct {
+		ID   string `json:"id"`
+		Type string `json:"type"`
+	}
+	if err := json.Unmarshal(raw, &envelope); err != nil {
+		return string(raw)
+	}
+	if id := strings.TrimSpace(envelope.ID); id != "" {
+		return id
+	}
+	return strings.TrimSpace(envelope.Type)
+}
+
+func cloneDispatchStatusSlice(transitions []DispatchStatus) []DispatchStatus {
+	if len(transitions) == 0 {
+		return nil
+	}
+	return append([]DispatchStatus(nil), transitions...)
+}
+
+func cloneProviderSessionRefs(refs []ProviderSessionRef) []ProviderSessionRef {
+	if len(refs) == 0 {
+		return nil
+	}
+	return append([]ProviderSessionRef(nil), refs...)
+}
+
+func cloneDispatchSummary(dispatch DispatchSummary) DispatchSummary {
+	cloned := dispatch
+	if len(dispatch.OutputArtifactIDs) > 0 {
+		cloned.OutputArtifactIDs = append([]string(nil), dispatch.OutputArtifactIDs...)
+	}
+	if len(dispatch.ProviderSessionRefs) > 0 {
+		cloned.ProviderSessionRefs = cloneProviderSessionRefs(dispatch.ProviderSessionRefs)
+	}
+	if dispatch.FailureDetail != nil {
+		detail := *dispatch.FailureDetail
+		cloned.FailureDetail = &detail
+	}
+	return cloned
 }
 
 // ProgressCounts summarizes durable dispatch progress for one session.
